auth/requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
autopep8==1.5.7
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.13.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
